backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
braintree==4.15.2
Brotli==1.0.9
celery==5.2.7
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cssselect2==0.6.0
Deprecated==1.2.13
dj-database-url==0.5.0
Django==4.0.5
django-admin-sortable2==2.0.4
django-dotenv==1.4.2
django-js-asset==2.0.0
django-mptt==0.13.4
django-stubs==1.11.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
easy-thumbnails==2.8.1
Faker==13.14.0
fonttools==4.33.3
gunicorn==20.1.0
html5lib==1.1
idna==3.3
kombu==5.2.4
lxml==4.9.0
mypy==0.950
mypy-extensions==0.4.3
packaging==21.3
Pillow==9.1.1
pkg_resources==0.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
pydyf==0.2.0
pyparsing==3.0.9
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
reportlab==3.6.10
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.3.0
tinycss2==1.1.1
tomli==2.0.1
types-pytz==2021.3.8
types-PyYAML==6.0.8
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
zopfli==0.2.1
